requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
python-dotenv==1.0.1
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.7
pandas